Feat : 알림 내용 부분 추가
알림 아이콘 추가 필요
